Revert "Feature/add edgedb"
