--- app/crawler/v3/dorm_crawler.py
@@ -1,102 +1,0 @@
-import json
-
-import requests
-from bs4 import BeautifulSoup
-import re
-from datetime import datetime
-import edgedb
-
-
-def dorm_parser(board: str, page: int):
-    client = edgedb.create_client()
-    now = datetime.now()
-
-    url = f"https://dorm.koreatech.ac.kr/content/board/list.php?now_page={page}&GUBN=&SEARCH=&BOARDID={board}"
-    response = requests.get(url, verify=False)
-
-    if response.status_code == 200:
-        html = response.text
-        soup = BeautifulSoup(html, 'html.parser')
-
-        posts = soup.select("#board > table > tbody > tr")
-        for post in posts:
-            try:
-                num_parsed = post.select_one("td:nth-child(1)").get_text().strip()
-                writer_parsed = post.select_one("td:nth-child(3)").get_text().strip()
-                write_date_parsed = post.select_one("td:nth-child(4)").get_text().strip()
-                write_date_parsed = datetime.strptime(write_date_parsed, '%Y-%m-%d')
-                if board == "notice":
-                    read_count_parsed = post.select_one("td:nth-child(6)").get_text().strip()
-                else:
-                    read_count_parsed = post.select_one("td:nth-child(5)").get_text().strip()
-                read_count_parsed = int(read_count_parsed)
-                article_url_parsed = post.select_one("td:nth-child(2) > a").get('href')
-                article_url_parsed = re.sub("&now_page=\d*", "", article_url_parsed)
-                article_url_parsed = f"https://dorm.koreatech.ac.kr/content/board/{article_url_parsed}"
-
-            except AttributeError:
-                continue
-
-            article_response = requests.get(article_url_parsed, verify=False)
-
-            if response.status_code == 200:
-                html = article_response.text
-                soup = BeautifulSoup(html, 'html.parser')
-                try:
-                    title_parsed = soup.select_one(
-                        "#board > div.boardViewer > h4").get_text().strip()
-                    text_parsed = soup.select_one(
-                        "#board > div.boardViewer > table.viewer div.story").decode_contents()
-                    text_parsed = text_parsed.replace("<img", "<br><img")
-
-                except AttributeError:
-                    continue
-
-                files = soup.select("#board > div.boardViewer > table.viewer tr:nth-child(3) > td > a")
-
-                file_list = []
-                for file in files:
-                    file_uri = file["href"]
-                    file_name = file.get_text()
-                    file_name = re.sub("\[.*]", "", file_name).strip()
-
-                    file_tuple = (file_name, file_uri)
-                    file_list.append(file_tuple)
-
-                try:
-                    client.query("""
-                        insert dorm {
-                            board := <str>$board,
-                            num := <str>$num,
-                            title := <str>$title,
-                            writer := <str>$writer,
-                            write_date := <cal::local_date>$write_date,
-                            read_count := <int64>$read_count,
-                            article_url := <str>$article_url,
-                            content := <str>$content,
-                            crawled_time := <cal::local_datetime>$crawled_time,
-                            files := <array<tuple<str, str>>><json>$files
-                        }
-                    """, board=board, num=num_parsed, title=title_parsed, writer=writer_parsed,
-                                 write_date=write_date_parsed, read_count=read_count_parsed,
-                                 article_url=article_url_parsed, content=text_parsed, crawled_time=now,
-                                 files=json.dumps(file_list))
-                except edgedb.errors.ConstraintViolationError:
-                    client.query("""
-                        update dorm
-                        filter .article_url = <str>$article_url
-                        set {
-                            title := <str>$title,
-                            write_date := <cal::local_date>$write_date,
-                            read_count := <int64>$read_count,
-                            content := <str>$content,
-                            crawled_time := <cal::local_datetime>$crawled_time,
-                            files := <array<tuple<str, str>>><json>$files
-                        }
-                    """, title=title_parsed, write_date=write_date_parsed, read_count=read_count_parsed,
-                                 content=text_parsed, crawled_time=now,
-                                 article_url=article_url_parsed, files=json.dumps(file_list))
-    else:
-        pass
-
-    client.close()
