--- edgedb.toml
@@ -1,2 +1,0 @@
-[edgedb]
-server-version = "2.14"
