+++ routers/v2/cse.py
@@ -1,8 +1,8 @@
 from fastapi import APIRouter, Depends
-from app.db.v3.cse import *
+from crawler.v2.cse_crawler import *
 
 router = APIRouter(
-    prefix="/v3/cse",
+    prefix="/v2/cse",
     tags=["cse"],
     responses={404: {"description": "Not found"}},
 )
@@ -29,5 +29,5 @@ async def get_cse_pds(params: dict = Depends(cse_pds)):
 
 
 @router.get("/article/")
-async def get_cse_article(params: dict = Depends(get_article)):
+async def get_cse_article(params: dict = Depends(cse_article_parser)):
     return params
